requirements.txt
deep-translator
nltk
rouge-score
